finish refactor of transfer estimation, and reorganize src and test
