+++ zeneth-js/src/fee-estimation.ts
@@ -7,25 +7,25 @@ import { formatEther, parseUnits } from '@ethersproject/units';
  * @notice Estimates the fee (in base token units) for a bundle
  * @param obj
  * @param obj.tokenAddress Token in which fee will be expressed as units of
- * @param obj.transferUpperBound Given upper bound gas fee (in wei) for transfer
- * @param obj.approveUpperBound Given upper bound gas fee (in wei) for approve
- * @param obj.swapUpperBound Given upper bound gas fee (in wei) for swap
+ * @param obj.gasLimitTransfer Given upper bound gas fee (in wei) for transfer
+ * @param obj.gasLimitApprove Given upper bound gas fee (in wei) for approve
+ * @param obj.gasLimitSwap Given upper bound gas fee (in wei) for swap
  * @param obj.flashbotsAdjustment Multiplier for flashbots bribing miner, e.g. 1.05 to add 5%
  * @returns Estimated fee for miner in token base units
  */
 export const estimateTransferFee = async ({
   tokenAddress,
   tokenDecimals,
-  transferUpperBound,
-  approveUpperBound,
-  swapUpperBound,
+  gasLimitTransfer,
+  gasLimitApprove,
+  gasLimitSwap,
   flashbotsAdjustment,
 }: {
   tokenAddress: string;
   tokenDecimals: number;
-  transferUpperBound: BigNumberish;
-  approveUpperBound: BigNumberish;
-  swapUpperBound: BigNumberish;
+  gasLimitTransfer: BigNumberish;
+  gasLimitApprove: BigNumberish;
+  gasLimitSwap: BigNumberish;
   flashbotsAdjustment: number;
 }): Promise<BigNumber> => {
   const [gasPriceInWei, tokenPrice, ethPrice] = await Promise.all([
@@ -34,7 +34,7 @@ export const estimateTransferFee = async ({
     getTokenPriceInUsd(ethAddress),
   ]);
 
-  const bundleGasUsed = BigNumber.from(transferUpperBound).add(approveUpperBound).add(swapUpperBound); // total gas needed
+  const bundleGasUsed = BigNumber.from(gasLimitTransfer).add(gasLimitApprove).add(gasLimitSwap); // total gas needed
   const bundlePriceInWei = bundleGasUsed.mul(gasPriceInWei); // total gas price in wei
   const scaledBundlePriceInWei = bundlePriceInWei.mul(BigNumber.from(flashbotsAdjustment * 1000)).div(1000); // total gas, scaled up
   const bundlePriceInEth = +formatEther(scaledBundlePriceInWei); // total gas, denominated in ETH
